Updated implementation of SwaggerSpec resource API class to better support the updated design.
This is an update in progress, resources and paths need to be generated yet, however this will be made much simpler through the use of the UrlPath class etc.

Operations are still missing some of the documentation required to produce a full swagger spec, this is however an implementation detail that I am working my way through.